software/requirements.txt
opencv-python>=4.8.0
ultralytics>=8.2.0
pyserial>=3.5
numpy>=1.24.0
